third_party/WebKit/LayoutTests/fast/css/csstext-of-content-string-expected.txt

Blocked access to external URL http://example.com/
Tests if a css text of content value is quoted when the value is string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.sheet.cssRules[0].cssText is "#A::after { content: \"A\"; }"
PASS style.sheet.cssRules[1].cssText is "#B::after { content: \"'B'\"; }"
PASS style.sheet.cssRules[2].cssText is "#C::after { content: \"'C\\8 '\"; }"
PASS style.sheet.cssRules[3].cssText is "#D::after { content: \"'D'\" url(\"http://example.com/\") \"EFG\"; }"
PASS successfullyParsed is true

TEST COMPLETE
